requirements.txt
Django==2.2.1
djangorestframework==3.9.3
Pygments==2.3.1
pytz==2019.1
sqlparse==0.3.0
